+++ lib/loot.h
@@ -12,8 +12,10 @@
 #define GROUND 1
 #define ROOF 2
 #define WALL 3
-#define DOOR 4
-#define NOITEM 5
+#define C_DOOR 4
+#define O_DOOR 5
+#define HOLDER 6
+#define NOITEM 7
 
 struct loot {
 	char *name;
@@ -29,7 +31,7 @@ extern char	 	*get_loot_name(short int id);
 extern char 		 is_loot_stackable(short int id);
 extern char		 get_loot_type(short int id);
 extern void		 pickup_item(struct game *cur_game, struct worldmap *map, struct player *cur_player);
-extern SDL_bool		 handle_pickup(struct worldmap *map, struct player *cur_player, int x, int y);
+extern SDL_bool		 handle_pickup(struct game *cur_game, struct worldmap *map, struct player *cur_player, int x, int y);
 extern void		 throw_item(struct game *cur_game, struct worldmap *map, struct player *cur_player);
 extern SDL_bool		 handle_throw(struct game *cur_game, struct worldmap *map, struct player *cur_player, int x, int y, short int quantity);
 extern void		 move_cursor(struct game *cur_game, char dir);