+++ lib/maps.h
@@ -14,7 +14,7 @@ struct worldmap {
 	int **tile;
 	short int **biome;
 	unsigned short int **loot;
-	unsigned char **quantity;
+	unsigned short int **quantity;
 	unsigned short int **ground;
 	unsigned short int **roof;
 	SDL_bool **harvestable;