+++ src/devm.c
@@ -31,9 +31,13 @@ void
 give_me_floors(struct player *cur_player)
 {
 	cur_player->loot[0] = 8;
-	cur_player->quantity[0] = 255;
+	cur_player->quantity[0] = 999;
 	cur_player->loot[1] = 9;
-	cur_player->quantity[1] = 255;
+	cur_player->quantity[1] = 999;
 	cur_player->loot[2] = 10;
-	cur_player->quantity[2] = 255;
+	cur_player->quantity[2] = 999;
+	cur_player->loot[3] = 11;
+	cur_player->quantity[3] = 10;
+	cur_player->loot[4] = 13;
+	cur_player->quantity[4] = 10;
 }